Move notification period helper to utils module
