+++ mtp_api/apps/notification/constants.py
@@ -1,6 +1,3 @@
-from datetime import datetime, timedelta, time
-
-from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from extended_choices import Choices
 
@@ -12,23 +9,3 @@ EMAIL_FREQUENCY = Choices(
     ('WEEKLY', 'weekly', _('Weekly')),
     ('MONTHLY', 'monthly', _('Monthly')),
 )
-
-
-def get_notification_period(email_frequency):
-    """
-    Returns start and end datetime of previous day, week, or month
-    """
-    today = timezone.make_aware(
-        datetime.combine(timezone.now(), time.min)
-    )
-    if email_frequency == EMAIL_FREQUENCY.DAILY:
-        return today - timedelta(days=1), today
-    elif email_frequency == EMAIL_FREQUENCY.WEEKLY:
-        start_of_this_week = today - timedelta(days=today.weekday())
-        return start_of_this_week - timedelta(days=7), start_of_this_week
-    elif email_frequency == EMAIL_FREQUENCY.MONTHLY:
-        first = today.replace(day=1)
-        return (first - timedelta(days=1)).replace(day=1), first
-    elif email_frequency == EMAIL_FREQUENCY.NEVER:
-        return today, today - timedelta(days=1)
-    raise KeyError
